requirements.txt
websockets==12.0
fastapi==0.103.1
Jinja2==3.1.2
uvicorn==0.23.2
aiortc==1.5.0
python-dotenv==1.0.0